+++ target/modest_client/handlers.h
@@ -24,6 +24,8 @@ ETERM *
 handle_insert_after(ErlMessage* emsg, ETERM *response);
 ETERM *
 handle_replace(ErlMessage* emsg, ETERM *response);
+ETERM *
+handle_slice(ErlMessage* emsg, ETERM *response);
 
 typedef struct _state_t {
   int fd;